+++ src/test/config.test.ts
@@ -30,6 +30,8 @@ const createEnvVars = (configPath: string): EnvironmentVariables => ({
 })
 
 const defaults = {
+  ACL_CONTEXT_BUILDER: undefined,
+  ACL_CONTEXT_BUILDER_PATH: '/usr/src/app/config/acl-context-builder.js',
   CONTENT_TYPE_MAP: defaultConfigs.CONTENT_TYPE_MAP,
   LANGUAGES_CONFIG: [],
   LANGUAGES_DIRECTORY_PATH: '/usr/static/languages',
@@ -86,6 +88,32 @@ describe('config injection tests', () => {
         },
       ],
       LANGUAGES_DIRECTORY_PATH: targetPath,
+      PUBLIC_HEADERS_MAP: {},
+      SERVICE_CONFIG_PATH: defaultConfigs.SERVICE_CONFIG_PATH,
+    })
+
+    await cleanup()
+  })
+
+  it('should parse a configuration with custom acl context builder', async () => {
+    const filename = 'acl-context-builder.js'
+    const { cleanup, name: targetPath } = await createTmpDir({
+      [filename]: 'export default () => { return [] }',
+    })
+    const aclContextBuilderPath = path.join(targetPath, filename)
+
+    const envVars = {
+      ...baseVariables,
+      ACL_CONTEXT_BUILDER_PATH: aclContextBuilderPath,
+    }
+
+    const config = parseConfig(envVars)
+    expect(config.ACL_CONTEXT_BUILDER).to.be.a('function')
+
+    config.ACL_CONTEXT_BUILDER = undefined
+    expect(config).to.deep.equal({
+      ...defaults,
+      ACL_CONTEXT_BUILDER_PATH: aclContextBuilderPath,
       PUBLIC_HEADERS_MAP: {},
       SERVICE_CONFIG_PATH: defaultConfigs.SERVICE_CONFIG_PATH,
     })